Change EffortModel signatures to return doubles rather than ints for effort values.

+++ src/main/model/effortmodel/EffortModel.java
@@ -12,8 +12,8 @@ public abstract class EffortModel {
     }
 
     // EFFECTS: computes the total effort of typing
-    public int computeTotalEffort() {
-        int totalEffort = 0;
+    public double computeTotalEffort() {
+        double totalEffort = 0.00;
 
         while (!corpus.isFinished()) {
             char key = corpus.consume();
@@ -24,6 +24,6 @@ public abstract class EffortModel {
         return totalEffort;
     }
 
-    public abstract int computePartialEffort(char key);
+    public abstract double computePartialEffort(char key);
 
 }
